cleaned up changes
