+++ authorizer.go
@@ -115,7 +115,7 @@ func (v *authorizer) Authorize() error {
 		successful := false
 		for _, query := range c.Queries {
 			// for checks that verify Preds with no Vars, len(res) is 0, there are no new facts!
-			res, err := v.world.QueryRule2(query, v.symbols)
+			res, err := v.world.QueryRule(query, v.symbols)
 			if len(*res) != 0 || err == nil {
 				successful = true
 				break
@@ -138,7 +138,7 @@ func (v *authorizer) Authorize() error {
 
 		successful := false
 		for _, query := range c.Queries {
-			res, err := v.world.QueryRule2(query, v.symbols)
+			res, err := v.world.QueryRule(query, v.symbols)
 			if len(*res) != 0 || err == nil {
 				successful = true
 				break
@@ -160,7 +160,7 @@ func (v *authorizer) Authorize() error {
 		}
 		for _, query := range policy.Queries {
 			// TODO: how to deal with err? len(res) can be 0 for checks with pred that don't have Vars
-			res, err := v.world.QueryRule2(query.convert(v.symbols), v.symbols)
+			res, err := v.world.QueryRule(query.convert(v.symbols), v.symbols)
 			if len(*res) != 0 || err == nil {
 				switch policy.Kind {
 				case PolicyKindAllow:
@@ -211,7 +211,7 @@ func (v *authorizer) Authorize() error {
 
 			successful := false
 			for _, query := range c.Queries {
-				res, err := block_world.QueryRule2(query, v.symbols)
+				res, err := block_world.QueryRule(query, v.symbols)
 
 				if len(*res) != 0 || err == nil {
 					successful = true
@@ -255,7 +255,7 @@ func (v *authorizer) Query(rule Rule) (FactSet, error) {
 	}
 	v.dirty = true
 
-	facts, err := v.world.QueryRule2(rule.convert(v.symbols), v.symbols)
+	facts, err := v.world.QueryRule(rule.convert(v.symbols), v.symbols)
 	if err != nil {
 		return nil, err
 	}
