+++ datalog/datalog_test.go
@@ -43,7 +43,7 @@ func TestFamily(t *testing.T) {
 	}
 
 	t.Logf("querying r1: %s", dbg.Rule(r1))
-	queryRuleResult, _ := w.QueryRule2(r1, syms)
+	queryRuleResult, _ := w.QueryRule(r1, syms)
 	t.Logf("r1 query: %s", dbg.FactSet(queryRuleResult))
 	t.Logf("current facts: %s", dbg.FactSet(w.facts))
 
@@ -105,7 +105,7 @@ func TestNumbers(t *testing.T) {
 	w.AddFact(Fact{Predicate{t2, []Term{Integer(1), bbb, Integer(0)}}})
 	w.AddFact(Fact{Predicate{t2, []Term{Integer(2), ccc, Integer(1)}}})
 
-	res, _ := w.QueryRule2(Rule{
+	res, _ := w.QueryRule(Rule{
 		Head: Predicate{join, []Term{hashVar("left"), hashVar("right")}},
 		Body: []Predicate{
 			{t1, []Term{hashVar("id"), hashVar("left")}},
@@ -121,7 +121,7 @@ func TestNumbers(t *testing.T) {
 		t.Errorf("query failed:\n have: %s\n want: %s", dbg.FactSet(res), dbg.FactSet(expected))
 	}
 
-	res, _ = w.QueryRule2(Rule{
+	res, _ = w.QueryRule(Rule{
 		Head: Predicate{join, []Term{hashVar("left"), hashVar("right")}},
 		Body: []Predicate{
 			{t1, []Term{Variable(1234), hashVar("left")}},
@@ -160,7 +160,7 @@ func TestString(t *testing.T) {
 	w.AddFact(Fact{Predicate{route, []Term{Integer(4), app1, syms.Insert("mx.example.com")}}})
 
 	testSuffix := func(suffix string, syms *SymbolTable) (*FactSet, error) {
-		return w.QueryRule2(Rule{
+		return w.QueryRule(Rule{
 			Head: Predicate{suff, []Term{hashVar("app_id"), Variable(1234)}},
 			Body: []Predicate{{route, []Term{Variable(0), hashVar("app_id"), Variable(1234)}}},
 			Expressions: []Expression{{
@@ -206,7 +206,7 @@ func TestDate(t *testing.T) {
 	w.AddFact(Fact{Predicate{x, []Term{Date(t1.Unix()), abc}}})
 	w.AddFact(Fact{Predicate{x, []Term{Date(t3.Unix()), def}}})
 
-	res, _ := w.QueryRule2(Rule{
+	res, _ := w.QueryRule(Rule{
 		Head: Predicate{before, []Term{Variable(1234), hashVar("val")}},
 		Body: []Predicate{{x, []Term{Variable(1234), hashVar("val")}}},
 		Expressions: []Expression{{
@@ -224,7 +224,7 @@ func TestDate(t *testing.T) {
 		t.Errorf("before query failed:\n have: %s\n want: %s", dbg.FactSet(res), dbg.FactSet(expected))
 	}
 
-	res, _ = w.QueryRule2(Rule{
+	res, _ = w.QueryRule(Rule{
 		Head: Predicate{after, []Term{Variable(1234), hashVar("val")}},
 		Body: []Predicate{{x, []Term{Variable(1234), hashVar("val")}}},
 		Expressions: []Expression{{
@@ -277,7 +277,7 @@ func TestBytes(t *testing.T) {
 	w.AddFact(Fact{Predicate{key, []Term{usr2, Bytes(k2)}}})
 	w.AddFact(Fact{Predicate{key, []Term{usr3, Bytes(k3)}}})
 
-	res, _ := w.QueryRule2(Rule{
+	res, _ := w.QueryRule(Rule{
 		Head: Predicate{keyMatch, []Term{hashVar("usr"), Variable(1)}},
 		Body: []Predicate{{key, []Term{hashVar("usr"), Variable(1)}}},
 		Expressions: []Expression{{
@@ -293,7 +293,7 @@ func TestBytes(t *testing.T) {
 		t.Errorf("key equal query failed:\n have: %s\n want: %s", dbg.FactSet(res), dbg.FactSet(expected))
 	}
 
-	res, _ = w.QueryRule2(Rule{
+	res, _ = w.QueryRule(Rule{
 		Head: Predicate{keyMatch, []Term{hashVar("usr"), Variable(1)}},
 		Body: []Predicate{{key, []Term{hashVar("usr"), Variable(1)}}},
 		Expressions: []Expression{{
@@ -310,7 +310,7 @@ func TestBytes(t *testing.T) {
 		t.Errorf("key in query failed:\n have: %s\n want: %s", dbg.FactSet(res), dbg.FactSet(expected))
 	}
 
-	res, _ = w.QueryRule2(Rule{
+	res, _ = w.QueryRule(Rule{
 		Head: Predicate{keyMatch, []Term{hashVar("usr"), Variable(1)}},
 		Body: []Predicate{{key, []Term{hashVar("usr"), Variable(1)}}},
 		Expressions: []Expression{{
@@ -351,7 +351,7 @@ func TestResource(t *testing.T) {
 	w.AddFact(Fact{Predicate{right, []Term{authority, file1, write}}})
 
 	check1 := syms.Insert("check1")
-	res, _ := w.QueryRule2(Rule{
+	res, _ := w.QueryRule(Rule{
 		Head: Predicate{check1, []Term{file1}},
 		Body: []Predicate{{resource, []Term{ambient, file1}}},
 	}, syms)
@@ -370,7 +370,7 @@ func TestResource(t *testing.T) {
 		},
 	}
 	t.Logf("r2 = %s", dbg.Rule(r2))
-	res, _ = w.QueryRule2(r2, syms)
+	res, _ = w.QueryRule(r2, syms)
 	if len(*res) > 0 {
 		t.Errorf("unexpected facts: %s", dbg.FactSet(res))
 	}
