+++ datalog/debugger.go
@@ -1,0 +1,95 @@
+package datalog
+
+import (
+	"fmt"
+	"strings"
+)
+
+type SymbolDebugger struct {
+	*SymbolTable
+}
+
+func (d SymbolDebugger) Predicate(p Predicate) string {
+	strs := make([]string, len(p.Terms))
+	for i, id := range p.Terms {
+		var s string
+		if sym, ok := id.(String); ok {
+			s = "\"" + d.Str(sym) + "\""
+		} else if variable, ok := id.(Variable); ok {
+			s = "$" + d.Var(variable)
+		} else {
+			s = fmt.Sprintf("%v", id)
+		}
+		strs[i] = s
+	}
+	return fmt.Sprintf("%s(%s)", d.Str(p.Name), strings.Join(strs, ", "))
+}
+
+func (d SymbolDebugger) Rule(r Rule) string {
+	head := d.Predicate(r.Head)
+	preds := make([]string, len(r.Body))
+	for i, p := range r.Body {
+		preds[i] = d.Predicate(p)
+	}
+	expressions := make([]string, len(r.Expressions))
+	for i, e := range r.Expressions {
+		expressions[i] = d.Expression(e)
+	}
+
+	var expressionsStart string
+	if len(preds) > 0 && len(expressions) > 0 {
+		expressionsStart = ", "
+	}
+
+	return fmt.Sprintf("%s <- %s%s%s", head, strings.Join(preds, ", "), expressionsStart, strings.Join(expressions, ", "))
+}
+
+func (d SymbolDebugger) CheckQuery(r Rule) string {
+	preds := make([]string, len(r.Body))
+	for i, p := range r.Body {
+		preds[i] = d.Predicate(p)
+	}
+	expressions := make([]string, len(r.Expressions))
+	for i, e := range r.Expressions {
+		expressions[i] = d.Expression(e)
+	}
+
+	var expressionsStart string
+	if len(preds) > 0 && len(expressions) > 0 {
+		expressionsStart = ", "
+	}
+
+	return fmt.Sprintf("%s%s%s", strings.Join(preds, ", "), expressionsStart, strings.Join(expressions, ", "))
+}
+
+func (d SymbolDebugger) Expression(e Expression) string {
+	return e.Print(d.SymbolTable)
+}
+
+func (d SymbolDebugger) Check(c Check) string {
+	queries := make([]string, len(c.Queries))
+	for i, q := range c.Queries {
+		queries[i] = d.CheckQuery(q)
+	}
+	return fmt.Sprintf("check if %s", strings.Join(queries, " or "))
+}
+
+func (d SymbolDebugger) World(w *World) string {
+	facts := make([]string, len(*w.facts))
+	for i, f := range *w.facts {
+		facts[i] = d.Predicate(f.Predicate)
+	}
+	rules := make([]string, len(w.rules))
+	for i, r := range w.rules {
+		rules[i] = d.Rule(r)
+	}
+	return fmt.Sprintf("World {{\n\tfacts: %v\n\trules: %v\n}}", facts, rules)
+}
+
+func (d SymbolDebugger) FactSet(s *FactSet) string {
+	strs := make([]string, len(*s))
+	for i, f := range *s {
+		strs[i] = d.Predicate(f.Predicate)
+	}
+	return fmt.Sprintf("%v", strs)
+}
