+++ datalog/world.go
@@ -1,0 +1,191 @@
+package datalog
+
+import (
+	"context"
+	"errors"
+	"time"
+)
+
+type runLimits struct {
+	maxFacts      int
+	maxIterations int
+	maxDuration   time.Duration
+}
+
+var defaultRunLimits = runLimits{
+	maxFacts:      1000,
+	maxIterations: 100,
+	maxDuration:   2 * time.Millisecond,
+}
+
+var (
+	ErrWorldRunLimitMaxFacts      = errors.New("datalog: world runtime limit: too many facts")
+	ErrWorldRunLimitMaxIterations = errors.New("datalog: world runtime limit: too many iterations")
+	ErrWorldRunLimitTimeout       = errors.New("datalog: world runtime limit: timeout")
+)
+
+type WorldOption func(w *World)
+
+func WithMaxFacts(maxFacts int) WorldOption {
+	return func(w *World) {
+		w.runLimits.maxFacts = maxFacts
+	}
+}
+
+func WithMaxIterations(maxIterations int) WorldOption {
+	return func(w *World) {
+		w.runLimits.maxIterations = maxIterations
+	}
+}
+
+func WithMaxDuration(maxDuration time.Duration) WorldOption {
+	return func(w *World) {
+		w.runLimits.maxDuration = maxDuration
+	}
+}
+
+type World struct {
+	facts *FactSet
+	rules []Rule
+
+	runLimits runLimits
+}
+
+func NewWorld(opts ...WorldOption) *World {
+	w := &World{
+		facts:     &FactSet{},
+		runLimits: defaultRunLimits,
+	}
+
+	for _, opt := range opts {
+		opt(w)
+	}
+
+	return w
+}
+
+func (w *World) AddFact(f Fact) {
+	w.facts.Insert(f)
+}
+
+func (w *World) Facts() *FactSet {
+	return w.facts
+}
+
+func (w *World) AddRule(r Rule) {
+	w.rules = append(w.rules, r)
+}
+
+func (w *World) ResetRules() {
+	w.rules = make([]Rule, 0)
+}
+
+func (w *World) Rules() []Rule {
+	return w.rules
+}
+
+func (w *World) Run(syms *SymbolTable) error {
+	done := make(chan error)
+	ctx, cancel := context.WithTimeout(context.Background(), w.runLimits.maxDuration)
+	defer cancel()
+
+	go func() {
+		for i := 0; i < w.runLimits.maxIterations; i++ {
+			select {
+			case <-ctx.Done():
+				return
+			default:
+				var newFacts FactSet
+				for _, r := range w.rules {
+					select {
+					case <-ctx.Done():
+						return
+					default:
+						if err := r.Apply(w.facts, &newFacts, syms); err != nil {
+							if !errors.Is(SolvSafeForWorldError{}, err) {
+								done <- err
+								return
+							}
+						}
+					}
+				}
+
+				prevCount := len(*w.facts)
+				w.facts.InsertAll([]Fact(newFacts))
+
+				newCount := len(*w.facts)
+				if newCount >= w.runLimits.maxFacts {
+					done <- ErrWorldRunLimitMaxFacts
+					return
+				}
+
+				// last iteration did not generate any new facts, so we can stop here
+				if newCount == prevCount {
+					done <- nil
+					return
+				}
+			}
+		}
+		done <- ErrWorldRunLimitMaxIterations
+	}()
+
+	select {
+	case <-ctx.Done():
+		return ErrWorldRunLimitTimeout
+	case err := <-done:
+		return err
+	}
+}
+
+func (w *World) Query(pred Predicate) *FactSet {
+	res := &FactSet{}
+	for _, f := range *w.facts {
+		if f.Predicate.Name != pred.Name {
+			continue
+		}
+
+		// if the predicate has a different number of IDs
+		// the fact must not match
+		if len(f.Predicate.Terms) != len(pred.Terms) {
+			continue
+		}
+		/*minLen := len(f.Predicate.IDs)
+		if l := len(pred.IDs); l < minLen {
+			minLen = l
+		}*/
+
+		matches := true
+		for i := 0; i < len(pred.Terms); i++ {
+			fID := f.Predicate.Terms[i]
+			pID := pred.Terms[i]
+
+			if pID.Type() != TermTypeVariable {
+				if fID.Type() != pID.Type() || fID != pID {
+					matches = false
+					break
+				}
+
+			}
+		}
+
+		if matches {
+			res.Insert(f)
+		}
+	}
+	return res
+}
+
+func (w *World) QueryRule(rule Rule, syms *SymbolTable) (*FactSet, error) {
+	newFacts := &FactSet{}
+	return newFacts, rule.Apply(w.facts, newFacts, syms)
+}
+
+func (w *World) Clone() *World {
+	newFacts := new(FactSet)
+	*newFacts = *w.facts
+	return &World{
+		facts:     newFacts,
+		rules:     append([]Rule{}, w.rules...),
+		runLimits: w.runLimits,
+	}
+}
